+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## 0.3.0.1
 
-- []()
+- [#31](https://github.com/parsonsmatt/annotated-exception/pull/31)
     - The `Show` instance for `AnnotatedException` attempts to peek into the
       `SomeException` to provide a more useful exception type. Also gives the
       output of `show`.